% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_cost-eff_constructors.R
\name{branch_unif_params}
\alias{branch_unif_params}
\title{branch_unif_params constructor}
\usage{
branch_unif_params(pmin, pmax, name)
}
\arguments{
\item{pmin}{Minimum probability}

\item{pmax}{Maximum probability}

\item{name}{Node label}
}
\value{
list
}
\description{
Define properties and assign \code{branch_unif_params} class.
}
\examples{

}
