% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_QALY_tb.R
\name{calc_QALY_tb}
\alias{calc_QALY_tb}
\title{Calculate QALYs for active TB cases}
\usage{
calc_QALY_tb(intervals = NA, utility, age, start_delay = NA,
  discount_rate = 0.035, ...)
}
\arguments{
\item{intervals}{Time intervals for each utility}

\item{utility}{(list) Utility value of non-diseased individual e.g. 1. Utility value of diseased individual}

\item{age}{Ages in years; vector numeric}

\item{start_delay}{What time delay to time origin, to shift discounting to smaller values}

\item{discount_rate}{default 3.5\% per annum}

\item{...}{Additional arguments}
}
\value{
list of diseasefree, death, cured QALYs
}
\description{
Calculate the QALYs for each active TB individuals.
}
\details{
For each of 3 alternatives:
\itemize{
  \item diseasefree: to all-cause death
  \item fatality: case-fatality 12 months from notification
  \item cured: successfully treated for LTBI
 }

Assume that death if it happens is within the first year of active TB.
Assume that active TB cases when treated and survive first year are ~~fully cured~~.

Consider person-perspective (death) or NHS-perspective (exit uk)
by defining the particular time-to-event end point.
}
\examples{

}
