% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_ce.R
\name{ce_default}
\alias{ce_default}
\title{ce_default}
\usage{
ce_default(ce0, ce1)
}
\arguments{
\item{ce1}{}
}
\description{
Uses the first column of the status-quo matrices
for all status-quo comparisons.
}
\examples{

}
