% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_popmod_dectree_res.R
\name{combine_popmod_dectree_res}
\alias{combine_popmod_dectree_res}
\title{combine_popmod_dectree_res}
\usage{
combine_popmod_dectree_res(cohort, interv, popmod_res, dectree_res,
  folders = NA)
}
\arguments{
\item{cohort}{individual level data; dataframe}

\item{interv}{fixed model run inputs; list}

\item{popmod_res}{output of \code{activetb_qaly_cost()}; list}

\item{dectree_res}{output of \code{parallel_decision_tree()}; list}

\item{folders}{list of ouput folder locations}
}
\value{
list of cost-effective statistics:
\itemize{
  \item ce0: marginal status-quo. Costs and QALYs of each sim.
  \item ce1: marginal intervention. Costs and QALYs of each sim.
  \item ce_default: non-incremental cost-effectiveness i.e. dataframe with first column status-quo.
  \item ce_incr: incremental cost-effectivness i.e. dataframe with first column 0 and other screening cost minus status-quo.
  }
}
\description{
Combine cost and QALY outputs from decision tree model and population model
for overall cost-effectiveness samples.
}
\examples{

}
\seealso{
\code{\link{parallel_decision_tree}},
  \code{\link{activetb_qaly_cost}}
}
