% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costeff_stats.R
\name{costeff_stats}
\alias{costeff_stats}
\title{Cost-effectiveness Statistics}
\usage{
costeff_stats(scenario_dat, interv_QALY, interv_cost, pop_year)
}
\arguments{
\item{scenario_dat}{list}

\item{interv_QALY}{list of \code{scenario_QALY()} output}

\item{interv_cost}{list of \code{scenario_cost()} output}

\item{pop_year}{integer}
}
\value{
list
\itemize{
   \item QALY.statusquo
   \item QALY.screened
   \item E_cost_screened: mean average
   \item cost.screened_person
   \item cost.statusquo_person
   \item cost_incur
   \item cost.statusquo
   \item cost.screened
   \item E_QALY_screened: mean average
   \item QALY.screened_person
   \item QALY.statusquo_person
   \item QALYgain
   \item cost_incur_person
   \item E_cost_incur: mean average
   \item E_cost_incur_person: mean average
   \item QALYgain_person
   \item E_QALYgain: mean average
   \item E_QALYgain_person: mean average
}
}
\description{
For a scenario the population model with active TB cases.
}
\examples{

}
