% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_folders.R
\name{cp_in_data_to_out_dir}
\alias{cp_in_data_to_out_dir}
\title{Copy input data to output folder}
\usage{
cp_in_data_to_out_dir(file_names, to_dir)
}
\arguments{
\item{file_names}{vector of text strings}

\item{to_dir}{}
}
\description{
Copy input data to output folder
}
