% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_and_save_policies.R
\name{create_and_save_policies}
\alias{create_and_save_policies}
\title{create_and_save_policies}
\usage{
create_and_save_policies(incidence_list, endpoints, LTBI_test, treatment)
}
\arguments{
\item{incidence_list}{WHO incidence in country of origin groups to target screening}

\item{endpoints}{when to stop using costs and QALYs at time of exit or death}

\item{LTBI_test}{TSPOT or QFT}

\item{treatment}{6 months or 3 months LTBI regimens}
}
\value{
list of policies as inputs for the model.
}
\description{
Given the input argument creates a sort of grid array, i.e all permutations, version of an environment
which behaves like a list object.
}
\examples{

}
