% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_and_save_scenarios.R
\name{create_and_save_scenarios}
\alias{create_and_save_scenarios}
\title{Create and save scenarios}
\usage{
create_and_save_scenarios(file_tag)
}
\arguments{
\item{file_tag}{Trailing part of the Excel file name to identify specific sets of scenarios.}
}
\value{
none (save to project data folder)
}
\description{
Read in an Excel workbook consisting of cost and probability (p) sheets.
This is converted to a list of dataframes for model inputs
and a long flat array to easily inspect by eye. Saved to file.
}
\examples{

}
