% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision_tree_cluster.R
\name{decision_tree_cluster}
\alias{decision_tree_cluster}
\title{Decision tree}
\usage{
decision_tree_cluster(params, N.mc = 2,
  cost_dectree = "data/osNode_cost_2009.Rds",
  health_dectree = "data/osNode_health_2009.Rds", out_datatree = FALSE)
}
\arguments{
\item{params}{an element of a scenario list with probabilities and costs to substitue into decision tree; long format array}

\item{N.mc}{number of simulations; integer}

\item{cost_dectree}{data.tree saved as Rds file names (string); default to package folder}

\item{health_dectree}{data.tree saved as Rds file names (string); default to package folder}

\item{out_datatree}{Output full datatree object? This may be large. Will also save to csv for checking; default: FALSE}
}
\value{
list
\itemize{
  \item \code{mc_cost}: each simulation total expected cost
  \item \code{mc_health}: each simulation total expected QALY loss
  \item \code{subset_pop}: cohort population sizes and probabilities at specific node or groups of nodes.
  Specifically calculates for individuals with LTBI since these are the subset of particular interest in term of cure; dataframe headings are
  \itemize{
     \item LTBI_pre
     \item tests
     \item positive
     \item startTx
     \item completeTx
     \item cured
     \item LTBI_post
     \item p_LTBI_to_cured
     \item LTBI_tests
     \item LTBI_positive
     \item LTBI_startTx
     \item LTBI_completeTx
     }
  \item \code{osNode.cost}: data.tree object
  \item \code{osNode.health}: data.tree object
  \item \code{call}: original call with arguments
  \item \code{N.mc}: number of Monte-Carlo simulations
}
}
\description{
Calculate decision tree expected costs and QALY loss
for each simulation.
}
\examples{

}
