% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expected_cost_QALY.R
\name{expected_cost_QALY}
\alias{expected_cost_QALY}
\title{expected_cost_QALY}
\usage{
expected_cost_QALY(cohort, means)
}
\arguments{
\item{means}{}
}
\description{
Calculate the average total cost using mean unit costs.
}
