% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy_cohort.R
\name{filter_cohort_by_policy}
\alias{filter_cohort_by_policy}
\title{Create policy cohort}
\usage{
filter_cohort_by_policy(cohort_in, policy_name, interv)
}
\arguments{
\item{cohort_in}{total sample}

\item{interv}{list of conditions}
}
\value{
cohort
}
\description{
Filter individuals by policy definition.
}
