% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interv_constructor.R
\name{interv_constructor}
\alias{interv_constructor}
\title{interv_constructor}
\usage{
interv_constructor(N.mc = 1, cluster = FALSE, use_discount = TRUE,
  no_students = FALSE, force_everyone_stays = FALSE,
  screen_with_delay = TRUE, MAX_SCREEN_DELAY = 5, FUP_MAX_YEAR = 100,
  screen_age_range = 18:35, year_cohort = "2009",
  incidence_grps_screen = c("(0,50]", "(50,150]", "(150,250]",
  "(250,350]", "(350,1e+05]"), min_screen_length_of_stay = 0,
  ENDPOINT_cost = "death", ENDPOINT_QALY = "death")
}
\arguments{
\item{N.mc}{Global fixed constant; default 1}

\item{use_discount}{Global fixed constant; TRUE/FALSE.}

\item{screen_with_delay}{Rather than screen _everyone_ on entry screen at random 0-5 years from port of entry}

\item{FUP_MAX_YEAR}{Time horizon for active TB progression; default 100}

\item{year_cohort}{2012 is most recent complete year; largest cohort, corresponds with Pareek () LTBI risk}

\item{incidence_grps_screen}{Modified in the deterministic sensitivity analysis but set default values}

\item{min_screen_length_of_stay}{Modified in the deterministic sensitivity analysis but set default values}

\item{ENDPOINT_cost}{Modified in the deterministic sensitivity analysis but set default values; exit uk or death}

\item{ENDPOINT_QALY}{Modified in the deterministic sensitivity analysis but set default values; exit uk or death}
}
\description{
Simple list making function with defaults.
}
