% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_to_BCEA.R
\name{list_to_BCEA}
\alias{list_to_BCEA}
\alias{list_to_BCEA_incr}
\title{list_to_BCEA}
\usage{
list_to_BCEA(scenario_list, discount = 1)

list_to_BCEA_incr(scenario_list, discount = 1)
}
\arguments{
\item{discount}{}
}
\description{
transform to BCEA package input format
}
\examples{

}
