% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_incremental_ce.R
\name{make_incremental_ce}
\alias{make_incremental_ce}
\title{make_incremental_ce}
\usage{
make_incremental_ce(popmod_res, t_dectree, sdiscount, folders = NA)
}
\arguments{
\item{popmod_res}{\code{activetb_qaly_cost()} output}

\item{t_dectree}{list of \code{mc_cost} and \code{mc_health} from \code{parallel_decision_tree()} output}

\item{sdiscount}{average discounting due to delay to starting screening}

\item{folders}{list of string locations}
}
\value{
list of incremental e and c
}
\description{
For plotting purposed in particular, we want the cost and QALY
output data to be in the same format as used by the \code{BCEA} package.
}
\examples{

}
