% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expected_cost_QALY.R
\name{mean_QALYs}
\alias{mean_QALYs}
\title{mean_QALYs}
\usage{
mean_QALYs(cohort, p_contact_tracing)
}
\arguments{
\item{p_contact_tracing}{}
}
\value{
list status-quo and disease-free expected QALY
}
\description{
Contact tracing
}
