% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myToDataFrame_fns.R
\name{my_ToDataFrameTable}
\alias{my_ToDataFrameTable}
\title{my_ToDataFrameTable}
\usage{
my_ToDataFrameTable(x, ..., pruneFun = NULL)
}
\arguments{
\item{pruneFun}{}
}
\description{
This is the same as the same named function in \code{data.tree}
except it is not filtered by leaf.
}
\seealso{
\link[data.tree]{ToDataFrameTable}
}
