% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myToDataFrame_fns.R
\name{my_ToDataFrameTypeCol}
\alias{my_ToDataFrameTypeCol}
\title{my_ToDataFrameTypeCol}
\usage{
my_ToDataFrameTypeCol(x, ..., type = "level", prefix = type,
  pruneFun = NULL)
}
\arguments{
\item{pruneFun}{}
}
\description{
This is the same as the same named function in \code{data.tree}
except it is not filtered by leaf.
}
\seealso{
\link[data.tree]{ToDataFrameTypeCol}
}
