% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_cost.R
\name{notif_cost}
\alias{notif_cost}
\title{Combined cost for each TB case}
\usage{
notif_cost(cost, probs, num_contacts, discounts)
}
\arguments{
\item{cost}{vector}

\item{probs}{vector}

\item{num_contacts}{vector, per case}

\item{discounts}{at time of notification; vector, per case}
}
\description{
including secondary infections. with discounting
}
\examples{

}
