% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_decision_tree.R
\name{parallel_decision_tree}
\alias{parallel_decision_tree}
\title{Parallel cost-effectiveness decision tree}
\usage{
parallel_decision_tree(scenario_params, interv, folders,
  out_datatree = FALSE)
}
\arguments{
\item{scenario_params}{list of dataframes of cost and probability values}

\item{interv}{list of fixed policy parameters}

\item{folders}{list os strings of locations for data and plots}

\item{out_datatree}{include in return and save data.tree object; default: FALSE}
}
\value{
List of \code{decision_tree_cluster} outputs for each scenario
}
\description{
Based on code from here:
\url{https://www.r-bloggers.com/how-to-go-parallel-in-r-basics-tips/}
}
\examples{

}
