% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_cost.R
\name{rcontact_tracing_costs}
\alias{rcontact_tracing_costs}
\title{randomly sample constact tracing costs}
\usage{
rcontact_tracing_costs(unit_cost)
}
\arguments{
\item{unit_cost}{named list}
}
\value{
named vector matching \code{p_contact_tracing}:
contact, aTB_Dx, aTB_Tx, LTBI_DxTx, index
}
\description{
randomly sample constact tracing costs
}
\examples{

}
