% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rows_first_n_ids.R
\name{rows_first_n_ids}
\alias{rows_first_n_ids}
\title{rows_first_n_ids}
\usage{
rows_first_n_ids(id_avoid, prop_avoid)
}
\arguments{
\item{id_avoid}{IDs, may have gaps/missing numbers}

\item{prop_avoid}{probability}
}
\value{
logical vector length ids
}
\description{
Finds the rows corresponding to the first \code{n} individuals
by ascending id numbers.
}
\seealso{
sample_avoid_lg, rsample_n_ids
}
