% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rows_first_n_ids.R
\name{rsample_n_ids}
\alias{rsample_n_ids}
\title{rsample_n_ids}
\usage{
rsample_n_ids(id_avoid, prop_avoid)
}
\arguments{
\item{prop_avoid}{probability}
}
\description{
rsample_n_ids
}
\seealso{
sample_avoid_lg, rows_first_n_ids
}
