% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_model.R
\name{run_model}
\alias{run_model}
\title{Run model}
\usage{
run_model(cohort_data = NA, make_plots = FALSE, sink_out = FALSE)
}
\arguments{
\item{cohort_data}{individual level data}

\item{make_plots}{TRUE/FALSE}

\item{sink_out}{output console to text file? Default: FALSE}
}
\description{
Wrapper around \code{\link{run_policy}}.
}
\examples{

}
\seealso{
\code{\link{run_policy}}
}
