% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_policy.R
\name{run_policy}
\alias{run_policy}
\title{run_policy}
\usage{
run_policy(cohort = NA, make_plots = FALSE)
}
\arguments{
\item{cohort}{individual level; default: NA}

\item{make_plots}{TRUE/FALSE}
}
\value{
empty
}
\description{
A single policy simulation (for multiple scenarios) of cost-effectiveness model.
}
\examples{

}
\seealso{
\code{\link{run_model}}
}
