% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rows_first_n_ids.R
\name{sample_avoid_lg}
\alias{sample_avoid_lg}
\title{sample_avoid_lg}
\usage{
sample_avoid_lg(id_avoided_tb, prop_avoided, ordered)
}
\arguments{
\item{prop_avoided}{probability}

\item{ordered}{random or ordered; TRUE/FALSE}
}
\description{
sample_avoid_lg
}
\seealso{
rows_first_n_ids, rsample_n_ids
}
