% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_subset_pop_dectree.R
\name{sample_subset_pop_dectree}
\alias{sample_subset_pop_dectree}
\title{sample_subset_pop_dectree}
\usage{
sample_subset_pop_dectree(osNode, n = 1, sample_p = TRUE)
}
\arguments{
\item{osNode}{data.tree object}

\item{n}{Sample size}

\item{sample_p}{Random sample TRUE/FALSE; default: TRUE}
}
\value{
matrix
}
\description{
Iteratively randomly samples probabilities
and then calculates subset sizes
}
\examples{

}
