% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_session_info.R
\name{save_session_info}
\alias{save_session_info}
\title{Save session info}
\usage{
save_session_info(file)
}
\arguments{
\item{file}{string location}
}
\description{
Text file includes package dependencies, versions, etc.
}
