% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_cost.R
\name{scenario_cost}
\alias{scenario_cost}
\title{Calculate total active TB cost of a scenario}
\usage{
scenario_cost(endpoint, unit_cost, probs_contact, cohort, prop_avoided,
  ordered = TRUE)
}
\arguments{
\item{endpoint}{\code{death} or \code{exit uk} i.e. time horizon}

\item{unit_cost}{Diagnosis and treatment cost distributions; list}

\item{probs_contact}{Proportions of individuals in subsets}

\item{cohort}{individual level dataframe. nrow total number of TB cases in EWNI and after exit}

\item{prop_avoided}{p_LTBI_to_cured; single numeric}

\item{order}{default TRUE}
}
\value{
list total cost for \code{statusquo} and \code{screened}; numeric
}
\description{
Calculate total active TB cost of a scenario
}
\examples{

}
