% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_QALY.R
\name{scenario_QALY}
\alias{scenario_QALY}
\title{Calculate total QALYs of a scenario}
\usage{
scenario_QALY(prop_avoided, endpoint, cohort, ordered = TRUE)
}
\arguments{
\item{prop_avoided}{probability cured of LTBI by screening}

\item{endpoint}{'death' or 'exit uk' for time horizon}

\item{cohort}{Individual level data}

\item{ordered}{Should individuals have a fixed order when avoiding TB
This is useful for reproducability and ensures that a higher proportion avoiding TB is always better; default: TRUE}
}
\value{
list of status-quo and screened total QALYs
}
\description{
Calculate total QALYs of a scenario
}
\examples{

}
