% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_QALYloss.R
\name{scenario_QALYloss}
\alias{scenario_QALYloss}
\title{scenario_QALYloss}
\usage{
scenario_QALYloss(prop_avoided, endpoint, cohort)
}
\arguments{
\item{prop_avoided}{proportion LTBI cured via screening}

\item{endpoint}{exit uk or death time horizon}

\item{cohort}{individual level data}
}
\value{
list \itemize{
  \item statusquo_mortality
 \item statusquo_morbidity
 \item screened_mortality
 \item screened_morbidity
 \item statusquo_mort_pp
 \item statusquo_morb_pp
 \item screened_mort_pp
 \item screened_morb_pp
 }
}
\description{
Splits output also into due to morbidity and mortality.
}
\examples{

}
