% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen_discount.R
\name{screen_discount}
\alias{screen_discount}
\title{screen_discount}
\usage{
screen_discount(cohort, discount_rate = 0.035)
}
\arguments{
\item{cohort}{individual level data}

\item{discount_rate}{default: 3.5\% per annum}
}
\value{
single numeric proportion between 0 and 1
}
\description{
Average discount cost and QALYs in decision tree
due to delayed start of screening from port of entry.
}
\examples{

}
