% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screened_cohort_cost.R
\name{screened_cohort_cost}
\alias{screened_cohort_cost}
\title{Calculate Potentially Screened Cohort Costs}
\usage{
screened_cohort_cost(n.diseasefree, cost.statusquo, unit_cost_case)
}
\arguments{
\item{n.diseasefree}{Number of disease-free individuals}

\item{cost.statusquo}{Cost under status-quo}

\item{unit_cost_case}{Unit cost of detect and treat an active TB case}
}
\value{
Total cost for potentially screened cohort
}
\description{
Substract the avoided cost of those successfully screened
from status-quo cost.
}
\examples{

}
\seealso{
\code{\link{screened_cohort_QALYs}}
}
