% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_branch_uniform_params.R
\name{set_branch_uniform_params}
\alias{set_branch_uniform_params}
\alias{set_branch_uniform_params.default}
\alias{set_branch_uniform_params.branch_unif_params}
\alias{set_branch_uniform_params.test}
\title{set_branch_uniform_params}
\usage{
set_branch_uniform_params(vals, osNode)

\method{set_branch_uniform_params}{default}(vals, osNode)

\method{set_branch_uniform_params}{branch_unif_params}(vals, osNode)

\method{set_branch_uniform_params}{test}(vals, osNode)
}
\arguments{
\item{osNode}{}
}
\description{
set_branch_uniform_params
}
\examples{

}
