% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy_interv.R
\name{set_policy}
\alias{set_policy}
\title{set_policy}
\usage{
set_policy(policy_name, interv)
}
\arguments{
\item{policy_name}{string}

\item{interv}{list of fixed model parameter values}
}
\description{
Set the intervention parameter values within an environment.
}
\examples{

}
