% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_folders.R
\name{setup_folders}
\alias{setup_folders}
\title{Setup folders}
\usage{
setup_folders(policy_name, interv)
}
\arguments{
\item{policy_name}{String}

\item{interv}{List of model run constants}
}
\value{
List of folder locations
}
\description{
Setup folders
}
\examples{

}
