% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_pop_dectree.R
\name{subset_pop_dectree}
\alias{subset_pop_dectree}
\title{Subset populations of decision tree}
\usage{
subset_pop_dectree(osNode)
}
\arguments{
\item{osNode}{data.tree object}
}
\value{
data.frame of probabilities
}
\description{
Specific to the LTBI screening model,
this gives the total probabilities of
particular state on the pathway by summing
across nodes, using pathprobs.
}
\examples{

}
