% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_cost-eff_constructors.R
\name{test}
\alias{test}
\title{TB test constructor}
\usage{
test(sens, spec)
}
\arguments{
\item{sens}{test sensitivity}

\item{spec}{test specificity}
}
\value{
list
}
\description{
Define properties and assign \code{test} class.
}
\examples{

}
