% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_cost.R
\name{total_contact_tracing_cost}
\alias{total_contact_tracing_cost}
\title{total_contact_tracing_cost}
\usage{
total_contact_tracing_cost(num_contacts, cost, probs)
}
\arguments{
\item{num_contacts}{vector, per case}

\item{probs}{vector}

\item{costs}{vector}
}
\value{
vector, per case
}
\description{
For an index case.
}
\examples{

}
